+++ tests/modules/test_Grinder.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import re
 import shutil
 import time
 
@@ -73,8 +74,6 @@ class TestGrinderExecutor(BZTestCase):
         obj.engine.provisioning = Local()
         obj.engine.provisioning.engine = obj.engine
         obj.engine.provisioning.executors = [obj]
-        obj.engine.provisioning.post_process()
-        self.assertTrue(isinstance(obj.engine.stopping_reason, RuntimeWarning))
 
     def test_with_results(self):
         obj = GrinderExecutor()
@@ -124,6 +123,61 @@ class TestGrinderExecutor(BZTestCase):
         finally:
             obj.shutdown()
         obj.post_process()
+        self.assertFalse(obj.has_results())
+
+    def test_script_generation(self):
+        obj = GrinderExecutor()
+        obj.engine = EngineEmul()
+        obj.settings.merge({'path': __dir__() + "/../grinder/fake_grinder.jar"})
+        obj.execution.merge({
+            "scenario": {
+                "default-address": "http://blazedemo.com",
+                "headers": {
+                    "My-Header": "Its-Value",
+                    "Another-Header": "Another-Value",
+                },
+                "timeout": "30s",
+                "think-time": "2s",
+                "store-cookie": True,
+                "requests": [
+                    '/',
+                    {'url': 'http://example.com/',
+                     'method': 'POST',
+                     'think-time': "1s",
+                     'headers': {
+                         'Custom': 'Header',
+                     }},
+                ]
+            }
+        })
+        obj.prepare()
+        script = open(os.path.join(obj.engine.artifacts_dir, 'grinder_requests.py')).read()
+
+        default_addr = re.findall(r"url='http://blazedemo.com'", script)
+        self.assertEquals(1, len(default_addr))
+
+        requests = re.findall(r"request\.([A-Z]+)\('(.+?)'", script)
+        self.assertEquals(2, len(requests))
+        self.assertEquals(requests[0], ('GET', '/'))
+        self.assertEquals(requests[1], ('POST', 'http://example.com/'))
+
+        sleeps = re.findall(r"grinder\.sleep\((.+)\)", script)
+        self.assertEquals(2, len(sleeps))
+        self.assertEquals(sleeps[0], '2000')
+        self.assertEquals(sleeps[1], '1000')
+
+        headers = re.findall(r"NVPair\('(.+)', '(.+)'\)", script)
+        self.assertEquals(3, len(headers))
+        self.assertIn(("My-Header", "Its-Value"), headers)
+        self.assertIn(("Another-Header", "Another-Value"), headers)
+        self.assertIn(("Custom", "Header"), headers)
+
+        timeout = re.findall(r"defaults.setTimeout\((\d+)\)", script)
+        self.assertEquals(1, len(timeout))
+        self.assertEquals(timeout[0], '30000')
+
+        cookies = re.findall(r"defaults.setUseCookies\(1\)", script)
+        self.assertEquals(1, len(cookies))
 
 
 class TestDataLogReader(BZTestCase):