+++ site/dat/docs/Changelog.md
@@ -2,6 +2,9 @@
 
 ## 1.7.1 (upcoming)
  - fix Mocha runner race condition
+ - support more scenario options for Grinder script generation
+ - add move check for executor results availability to provisioning
+ - add common `run-at` option for services
 
 ## 1.7.0 <sup>2 oct 2016</sup>
  - add RSpec tests runner for Selenium